+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * `Object.checksum(mixed)` will now treat Dates & Numbers differently
 * `Object.checksum(mixed)` will now correctly hash node Buffers
+* Add `Buffer` support to `Object.alike(a, b)`
 
 ## 0.5.10 (2018-07-11)